Move ESEF 2022 impl into ESEF plugin subdirectory
