+++ arelle/plugin/validate/ESEF/ESEF_Current/Image.py
